Update main.cpp
Removed the instance where Board.jpg was being displayed right after menu closed. There is already a function above main for this. I also remove the delete and set menu to null ptr to the end of the function as it caused an error when allocating memory for val. I also put everything we have so far under the conditional checking for val. This checks if the player chose to play or quit on the menu.